+++ src/main/java/sirpotato/Command.java
@@ -1,3 +1,5 @@
+package sirpotato;
+
 import java.io.IOException;
 
 public abstract class Command {
