+++ src/main/java/sirpotato/Deadline.java
@@ -1,3 +1,5 @@
+package sirpotato;
+
 import java.time.LocalDate;
 import java.time.format.DateTimeFormatter;
