+++ src/main/java/sirpotato/DeleteCommand.java
@@ -1,14 +1,16 @@
-public class MarkCommand extends Command {
+package sirpotato;
+
+public class DeleteCommand extends Command {
 
     private int itemNumber;
 
-    public MarkCommand(int itemNumber) {
+    public DeleteCommand(int itemNumber) {
         this.itemNumber = itemNumber;
     }
 
     @Override
     public void execute(TaskList tasks, Ui ui, Storage storage) {
-        tasks.mark(itemNumber);
+        tasks.delete(itemNumber, tasks.getList());
     }
     
 }
