+++ src/main/java/sirpotato/DukeException.java
@@ -1,10 +1,11 @@
+package sirpotato;
+
 /**
  * This is the DukeException
  * for when the user types incorrect inputs
  * such as forgetting the description of the task
  * or forgetting to put the /by
  */
-
 public class DukeException extends Exception {
     
     public DukeException(String message) {
