+++ src/main/java/sirpotato/Event.java
@@ -1,3 +1,5 @@
+package sirpotato;
+
 import java.time.LocalDate;
 import java.time.format.DateTimeFormatter;
