+++ src/main/java/sirpotato/Parser.java
@@ -1,3 +1,5 @@
+package sirpotato;
+
 import java.time.format.DateTimeFormatter;
 import java.time.LocalDate;
 import java.time.format.DateTimeParseException;
