+++ src/main/java/sirpotato/SirPotato.java
@@ -1,9 +1,13 @@
+package sirpotato;
+
 import java.util.Scanner;  // For user inputs
 import java.util.ArrayList; //To create the to-do list
+
 import java.io.File;
 import java.io.FileWriter;
 import java.io.FileNotFoundException;
 import java.io.IOException;
+
 import java.time.format.DateTimeFormatter;
 import java.time.LocalDate;
 import java.time.format.DateTimeParseException;
