+++ src/main/java/sirpotato/Storage.java
@@ -1,3 +1,5 @@
+package sirpotato;
+
 import java.util.Scanner; 
 import java.util.ArrayList; 
 import java.io.File;
