+++ src/main/java/sirpotato/Task.java
@@ -1,3 +1,5 @@
+package sirpotato;
+
 public class Task {
     protected String description;
     protected boolean isDone;
