+++ src/main/java/sirpotato/TaskList.java
@@ -1,3 +1,5 @@
+package sirpotato;
+
 import java.util.ArrayList;
 
 public class TaskList {
