+++ src/main/java/sirpotato/Todo.java
@@ -1,3 +1,5 @@
+package sirpotato;
+
 public class Todo extends Task {
 
     public Todo(String description) {
