+++ src/main/java/sirpotato/Ui.java
@@ -1,3 +1,5 @@
+package sirpotato;
+
 import java.util.ArrayList;
 import java.util.Scanner;
